@@ -8,7 +8,7 @@ use ParallelTask\Task\TaskRunner;
 use ParallelTask\Task\TaskRunnerNullSupervisor;
 use ParallelTask\Task\TaskScheduler;
 
-class FlowTest extends \PHPUnit_Framework_TestCase
+class FlowTest extends \PHPUnit\Framework\TestCase
 {
     public function testFlowSuccess()
     {
@@ -47,7 +47,7 @@ class FlowTest extends \PHPUnit_Framework_TestCase
 
         $futureResult = $executor->submit('testType', TestTask::class, $parameters);
         $taskRunner->runOnce('testType');
-        $this->setExpectedException(\Exception::class);
+        $this->expectException(\Exception::class);
         $futureResult->getResult();
     }
 }
